[server/refactor] korean-text-send : grpc 단에 선언 되어있던 AiClient utils 로 옮겨 websocket 에서도 사용 가능하게 수정
